Created CollapsibleLegend.
