+++ domains/rem/src/ui/generatedDates/GeneratedDates.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { __ } from '@wordpress/i18n';
 
-import { Heading, Legend } from '@eventespresso/components';
+import { Legend } from '@eventespresso/components';
 
 import Datetimes from './Datetimes';
 import useGenerateDates from './useGenerateDates';
@@ -18,9 +18,6 @@ const GeneratedDates: React.FC = () => {
 					<Datetimes datetimes={datetimes} />
 				</div>
 			</div>
-			<Heading as='h6' topBordered>
-				{__('Legend')}
-			</Heading>
 			<Legend direction='row' legendConfig={legendConfig} />
 		</>
 	);
