+++ packages/components/src/Collapsible/index.tsx
@@ -4,12 +4,7 @@ import { useSpring, animated } from 'react-spring';
 import type { CollapsibleProps } from './types';
 import { useRect } from '@eventespresso/hooks';
 
-/**
- * Collapsible
- * This component is an extraction from EntityListFilterBar
- * and is intended to be used only in context with that parent component.
- */
-const Collapsible: React.FC<CollapsibleProps> = ({ children, show = false }) => {
+export const Collapsible: React.FC<CollapsibleProps> = ({ children, className, show = false }) => {
 	const ref = useRef<HTMLDivElement>();
 	const { height } = useRect(ref);
 	const props = useSpring({
@@ -19,11 +14,9 @@ const Collapsible: React.FC<CollapsibleProps> = ({ children, show = false }) => 
 
 	return (
 		<animated.div style={props}>
-			<div className='ee-filter-bar__collapsible' ref={ref}>
+			<div className={className} ref={ref}>
 				{show && children}
 			</div>
 		</animated.div>
 	);
 };
-
-export default Collapsible;
