+++ packages/components/src/Collapsible/types.ts
@@ -1,0 +1,4 @@
+export interface CollapsibleProps {
+	className?: string;
+	show: boolean;
+}
