+++ packages/components/src/EntityList/filterBar/EntityListFilterBar.tsx
@@ -1,15 +1,13 @@
 import React, { useEffect, useState, useCallback } from 'react';
 import { __ } from '@wordpress/i18n';
 
-import Collapsible from './Collapsible';
-import type { EntityListFilterBarProps } from './types';
-import { Legend, SearchInput } from '../../..';
-import ToggleLegendButton from './buttons/ToggleLegendButton';
+import { Collapsible, Legend, SearchInput, ToggleLegendButton } from '../../..';
 import ToggleFiltersButton from './buttons/ToggleFiltersButton';
 import ToggleSortingButton from './buttons/ToggleSortingButton';
 import EntityListViewButtonGroup from './buttons/EntityListViewButtonGroup';
 import { useFilterBarUIElements } from '@eventespresso/registry';
 import { EntityListFilterStateManager as ELFSM } from './filterState';
+import type { EntityListFilterBarProps } from './types';
 
 import './style.scss';
 
@@ -57,14 +55,14 @@ const EntityListFilterBar = <FS extends ELFSM>({
 				/>
 				<ToggleSortingButton listId={listId} sortingEnabled={sortingEnabled} toggleSorting={toggleSorting} />
 				<ToggleLegendButton
-					listId={listId}
+					className='ee-filter-bar__btn ee-btn--small'
+					isDisabled={sortingEnabled}
 					showLegend={showLegend}
 					toggleLegend={toggleLegend}
-					isDisabled={sortingEnabled}
 				/>
 			</div>
 
-			<Collapsible show={showEntityFilters}>
+			<Collapsible className='ee-filter-bar__collapsible' show={showEntityFilters}>
 				{filerBarItems}
 				<div className='ee-filter-bar__filter'>
 					<SearchInput
@@ -77,7 +75,7 @@ const EntityListFilterBar = <FS extends ELFSM>({
 				</div>
 			</Collapsible>
 
-			<Collapsible show={showLegend}>
+			<Collapsible className='ee-filter-bar__collapsible' show={showLegend}>
 				<Legend legendConfig={legendConfig} />
 			</Collapsible>
 		</div>
