+++ packages/components/src/EntityList/filterBar/index.ts
@@ -1,5 +1,3 @@
-export * from './Collapsible';
-
 export { default as EntityListFilterBar } from './EntityListFilterBar';
 
 export * from './types';
