+++ packages/components/src/Legend/index.ts
@@ -1,3 +1,4 @@
 export { default as Legend } from './Legend';
+export { default as ToggleLegendButton } from './ToggleLegendButton';
 
 export * from './types';
